+++ assets/js/gmaps.js
@@ -5,6 +5,7 @@ function initialize() {
     var lng = jQuery(this).attr('data-lng');
     var zoom = parseInt(jQuery(this).attr('data-zoom'));
     var kml = jQuery(this).attr('data-kml');
+    var marker = jQuery(this).attr('data-marker');
     
     var myLatLng = new google.maps.LatLng(lat, lng);
     var mapOptions = {
@@ -14,11 +15,13 @@ function initialize() {
   
     var map = new google.maps.Map(jQuery(this)[0], mapOptions);
   
-    var marker = new google.maps.Marker({
-      position: myLatLng,
-      map: map,
-      title: 'Hello World!'
-    });
+    if (marker) {
+      var marker = new google.maps.Marker({
+        position: myLatLng,
+        map: map,
+        title: 'Hello World!'
+      });
+    }
     
     if (kml) {
       var ctaLayer = new google.maps.KmlLayer({