+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2
 
 - Add KML layer support
+- Add Marker activate/deactivate function
 
 # 0.1